admin/admin/requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.2.4
django-annoying==0.10.6
six==1.16.0
pika==1.2.0
psycopg2-binary==2.9.3
